refactor: rename modules to use 'Module' suffix for consistency and add new module implementations
